Moved DELETE persons/:d tests to persons.route.test
