+++ backend/src/routes/__test__/person.route.test.ts
@@ -3,7 +3,7 @@ import databaseOperations from '../../utils/test/db-handler';
 
 import Person, { PersonModel } from '../../models/person.model';
 import User, { UserModel } from '../../models/user.model';
-import { EncounterModel } from '../../models/encounter.model';
+import Encounter, { EncounterModel } from '../../models/encounter.model';
 import personService from '../../services/person.service';
 import app from '../../server';
 import "dotenv/config";
@@ -111,6 +111,15 @@ const encounter1Data: EncounterModel = {
   date: new Date('2022-02-23'),
   time_updated: new Date(Date.now()),
   description: 'Met at a cafe',
+  location: 'Auckland',
+  persons: [] as any
+}
+
+const encounter2Data: EncounterModel = {
+  title: "Encounter2",
+  date: new Date('2022-02-24'),
+  time_updated: new Date(Date.now()),
+  description: 'Had lunch together',
   location: 'Auckland',
   persons: [] as any
 }
@@ -460,3 +469,188 @@ describe('GET persons/:id', () => {
     expect(retrievedPerson.encounters[0].title).toEqual(encounter1Data.title)
   })
 });
+
+// Delete Person 200
+
+describe('DELETE /person/:id', () => {
+
+  it('Successfully deletes single person with no encounter: ', async () => {
+      // Get Authentication ID for User
+      const auth_id = await testUtils.getAuthIdFromToken(token);
+
+      // Create Person
+      const personOne = new Person(person1Data);
+      const personOneId = (await personOne.save())._id;
+
+      // Create User
+      const user = new User(user1Data);
+      // Add Encounter and Person ID to User encounters
+      user.persons.push(personOneId);
+      user.auth_id = auth_id;
+      await user.save();
+
+      await supertest(app).delete(`/api/persons/${personOneId}`)
+          .set('Accept', 'application/json')
+          .set('Authorization', token)
+          .expect(httpStatus.OK);
+
+      const newUser = await User.findOne({auth_id: user.auth_id});
+      expect(newUser?.persons).not.toContain(personOneId);
+  })
+
+  it('Successfully deletes single person with single encounter: ', async () => {
+      // Get Authentication ID for User
+      const auth_id = await testUtils.getAuthIdFromToken(token);
+
+      // Create Encounter
+      const encounterOne = new Encounter(encounter1Data);
+      const encounterOneId = (await encounterOne.save())._id;
+
+      // Create Person
+      const personOne = new Person(person1Data);
+      const personOneId = (await personOne.save())._id;
+
+      // Create User
+      const user = new User(user1Data);
+      // Add Encounter and Person ID to User encounters
+      user.encounters.push(encounterOneId);
+      user.persons.push(personOneId);
+      user.auth_id = auth_id;
+      await user.save();
+
+      // Add Encounter and Person IDs to respective objects
+      personOne.encounters.push(encounterOneId);
+      encounterOne.persons.push(personOneId);
+      await personOne.save();
+      await encounterOne.save();
+
+      await supertest(app).delete(`/api/persons/${personOneId}`)
+          .set('Accept', 'application/json')
+          .set('Authorization', token)
+          .expect(httpStatus.OK);
+
+      const newUser = await User.findOne({auth_id: user.auth_id});
+      expect(newUser?.persons).not.toContain(personOneId);
+  })
+
+  it('Successfully deletes multiple persons with multiple encounters', async () => {
+      // Get Authentication ID for User
+      const auth_id = await testUtils.getAuthIdFromToken(token);
+
+      // Create Encounter One
+      const encounterOne = new Encounter(encounter1Data);
+      const encounterOneId = (await encounterOne.save())._id;
+
+      // Create Encounter Two
+      const encounterTwo = new Encounter(encounter2Data);
+      const encounterTwoId = (await encounterTwo.save())._id;
+
+      // Create Person One
+      const personOne = new Person(person1Data);
+      const personOneId = (await personOne.save())._id;
+
+      // Create Person Two
+      const personTwo = new Person(person2Data);
+      const personTwoId = (await personTwo.save())._id;
+
+      // Create User
+      const user = new User(user1Data);
+      // Add Encounter and Person ID to User encounters
+      user.encounters.push(encounterOneId);
+      user.encounters.push(encounterTwoId);
+      user.persons.push(personOneId);
+      user.persons.push(personTwoId);
+      user.auth_id = auth_id;
+      await user.save();
+
+      // Add Encounter and Person IDs to respective objects
+      personOne.encounters.push(encounterOneId);
+      personOne.encounters.push(encounterTwoId);
+      encounterOne.persons.push(personOneId);
+      encounterOne.persons.push(personTwoId);
+
+      personTwo.encounters.push(encounterOneId);
+      personTwo.encounters.push(encounterTwoId);
+      encounterTwo.persons.push(personOneId);
+      encounterTwo.persons.push(personTwoId);
+
+      await personOne.save();
+      await personTwo.save();
+      await encounterOne.save();
+      await encounterTwo.save();
+
+      await supertest(app).delete(`/api/persons/${personOneId}`)
+          .set('Accept', 'application/json')
+          .set('Authorization', token)
+          .expect(httpStatus.OK);
+
+      await supertest(app).delete(`/api/persons/${personTwoId}`)
+          .set('Accept', 'application/json')
+          .set('Authorization', token)
+          .expect(httpStatus.OK);
+
+      // Check Encounters and Persons are removed
+      const newUser = await User.findOne({auth_id: user.auth_id});
+      expect(newUser?.persons).not.toContain(personOneId);
+      expect(newUser?.persons).not.toContain(personTwoId);
+
+      expect(newUser?.encounters).not.toContain(encounterOneId);
+      expect(newUser?.encounters).not.toContain(encounterTwoId);
+      
+      expect(await Person.findById({_id: personOneId})).toEqual(null);
+      expect(await Person.findById({_id: personTwoId})).toEqual(null);
+  })
+
+// Delete Person 404
+
+  it('Sends NOT_FOUND for invalid ID: ', async () => {
+      // Get Authentication ID for User
+      const auth_id = await testUtils.getAuthIdFromToken(token);
+
+      // Create Person
+      const personOne = new Person(person1Data);
+      const invalidPersonId = (await personOne.save())._id;
+
+      // Create User
+      const user = new User(user1Data);
+      // Add Encounter and Person ID to User encounters
+      user.auth_id = auth_id;
+      await user.save();
+
+      await supertest(app).delete(`/api/persons/${invalidPersonId}`)
+          .set('Accept', 'application/json')
+          .set('Authorization', token)
+          .expect(httpStatus.NOT_FOUND);
+
+      // Check that no encounters are deleted from User
+      const newUser = await User.findOne({auth_id: user.auth_id});
+      expect(newUser?.persons).toHaveLength(user.persons.length);
+  })
+
+// Delete Person 400
+  it('Sends BAD_REQUEST if Person ID is not in Collection: ', async () => {
+      // Get Authentication ID for User
+      const auth_id = await testUtils.getAuthIdFromToken(token);
+
+      // Create Person
+      const encounterOne = new Encounter(encounter1Data);
+      const invalidPersonId = (await encounterOne.save())._id;
+
+      // Create User
+      const user = new User(user1Data);
+      // Add invalid Person ID to User
+      user.auth_id = auth_id;
+      user.persons.push(invalidPersonId)
+      await user.save();
+
+      await supertest(app).delete(`/api/persons/${invalidPersonId}`)
+          .set('Accept', 'application/json')
+          .set('Authorization', token)
+          .expect(httpStatus.BAD_REQUEST);
+
+      // Check that invalid ID is deleted from User
+      const newUser = await User.findOne({auth_id: user.auth_id});
+      expect(newUser?.persons).not.toContain(invalidPersonId);
+  })
+})
+
